requirements.txt
argparse
requests
transformers
shodan
nmapthon2
sublist3r
